# Conflicts:
#	Networking/Networking/Remote/ProductsRemote.swift
#	WooCommerce/WooCommerce.xcodeproj/project.pbxproj
#	Yosemite/Yosemite/Stores/ProductStore.swift

+++ WooCommerce/Classes/ViewRelated/Products/Edit Product/Inventory Settings/ProductInventoryEditableData.swift
@@ -6,7 +6,7 @@ struct ProductInventoryEditableData: Equatable {
     let sku: String?
     let manageStock: Bool
     let soldIndividually: Bool?
-    let stockQuantity: Int64?
+    let stockQuantity: Decimal?
     let backordersSetting: ProductBackordersSetting?
     let stockStatus: ProductStockStatus?
 }
